scrappers/departures-scrapper/requirements.txt
kafka-python==2.0.2
avro==1.11.1
airportsdata==20230323

